feat: implement comprehensive unit test suite for Kafka protocol handlers and broker partition logic, and add CI/CD infrastructure.
