+++ .github/workflows/build_and_test.yml
@@ -1,0 +1,21 @@
+name: Go
+
+on:
+  push:
+    branches: ["main", "develop"]
+  pull_request:
+    branches: ["main"]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: actions/setup-go@v5
+        with:
+          go-version: "1.23"
+
+      - run: go build -v ./...
+
+      - run: go test -v -race ./...
